cli: add option to specify base image of a terminal command
Usually, the user needs to run a simple terminal command which needs a
specific set of libraries (and such requests often share the same set of
libraries). It would make sense that the client can install a Docker Hub
instance on one of their machines and use base images from there. But we
need to provide the user with a way to specify the base image for such
simple commands.